AIVN/requirements.txt
flask==2.3.3
google-generativeai==0.3.2
pillow==10.0.1
python-dotenv==1.0.0
pyyaml==6.0.1